+++ src/flutterwave-dotnet/APIs/Interfaces/ITransactions.cs
@@ -47,7 +47,7 @@ namespace Flutterwave.Net
         /// (This parameter is called "wait" on the offical docs)
         /// </param>
         /// <returns>Success</returns>
-        public ResendTransactionWebhookResponse ResendTransactionWebhook(int transactionId,
+        public ResendTransactionWebhookResponse ResendTransactionWebhook(long transactionId,
                                                                          bool shouldWait = false);
 
         /// <summary>
@@ -58,9 +58,9 @@ namespace Flutterwave.Net
         /// call as data.Id
         /// </param>
         /// <returns>The transaction with the specified id</returns>
-        public VerifyTransactionResponse VerifyTransaction(int transactionId);
+        public VerifyTransactionResponse VerifyTransaction(long transactionId);
 
-        public Task<VerifyTransactionResponse> VerifyTransactionAsync(int transactionId);
+        public Task<VerifyTransactionResponse> VerifyTransactionAsync(long transactionId);
 
         /// <summary>
         /// View transaction timeline
@@ -70,6 +70,6 @@ namespace Flutterwave.Net
         /// call as data.Id
         /// </param>
         /// <returns>The events for the transaction with the specified id</returns>
-        public ViewTransactionTimelineResponse ViewTransactionTimeline(int transactionId);
+        public ViewTransactionTimelineResponse ViewTransactionTimeline(long transactionId);
     }
 }
